Switched price from double to decimal
Added subtotal, tax and total calculations to shopping List View

+++ CommonLib/ExcelDataImporter.cs
@@ -52,7 +52,7 @@ namespace CommonLib
                         }
                         else if (col == 2)
                         {
-                            newItem.Price = double.Parse(temp);
+                            newItem.Price = decimal.Parse(temp);
                         }
                         else if (col == 3)
                         {
